Allow ply to read current kernel version
Add the -u option to use the current kernel version, rather than
requiring ply to be compiled for every different kernel. Of course
this introduces the possibility of issues with eBPF code mismatch,
but hopefully the verifier will do a good enough job to spot this
rather than just looking for a version mismatch.

Signed-off-by: icarus sparry <isparry+ply@gmail.com>